pyproject.toml
[project]
name = "events-face-service"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.5",
    "django-filter>=25.1",
    "django-rest-framework>=0.1.0",
]

[tool.ruff.lint]
extend-select = ["I"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[dependency-groups]
dev = [
    "ruff>=0.12.10",
]
